+++ src/main/java/com/minhcraft/mixin/block/CakeBlockMixin.java
@@ -5,17 +5,14 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.stats.Stats;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.CakeBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.gameevent.GameEvent;
-import net.minecraft.world.phys.BlockHitResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -56,6 +53,7 @@ public abstract class CakeBlockMixin {
                 }
                 player.addEffect(new MobEffectInstance(MobEffects.ABSORPTION, CAKE_ABSORPTION_DURATION, 1, false, false, false));
                 player.setAbsorptionAmount(Math.min(absorptionHearts+1, ModConfig.cakeMaxAbsorptionHearts));
+                level.playSound(player, player.blockPosition(), SoundEvents.GENERIC_EAT, SoundSource.PLAYERS, 1.0F, 1.0F);
                 level.playSound(player, player.blockPosition(), SoundEvents.PLAYER_BURP, SoundSource.PLAYERS, 1.0F, 1.0F);
             } else {
                 player.awardStat(Stats.EAT_CAKE_SLICE);